refactor Quantity, Require, and Sku classes for improved code clarity and functionality

+++ common-domain/src/main/java/com/factoryx/common/domain/Require.java
@@ -135,7 +135,7 @@ public final class Require {
     }
 
     public static int in(int value, int min, int max, String message) {
-        in((long) value, (long) min, (long) max, message);
+        in(value, min, (long) max, message);
         return value;
     }
